✨ feat: 평가 기준 설정하기 컴포넌트 분리
